refactor: ajust delete file function
