+++ libs/feature/src/lib/containers/index.ts
@@ -1,4 +1,4 @@
-export * from './files';
+export * from './file';
 export * from './login';
 export * from './protected-layout';
 export * from './test-page';
